requirements.txt
# AI Insight Reporter - Algorzen Research Division
# Python 3.10+ required

# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
seaborn>=0.12.0
matplotlib>=3.7.0

# PDF Generation
reportlab>=4.0.0

# Web UI
streamlit>=1.28.0

# AI Integration
openai>=1.3.0

# Configuration & Environment
python-dotenv>=1.0.0

# Optional Performance
polars>=0.19.0

# Table Formatting
tabulate>=0.9.0

# Utilities
Pillow>=10.0.0
scipy>=1.11.0
